test(ui): isolate assistant identity snapshot coverage
