+++ test/vitest/vitest.ui.config.ts
@@ -12,6 +12,7 @@ import { jsdomOptimizedDeps } from "./vitest.shared.config.ts";
 // keep this list in sync with vitest.ui-isolated.config.ts's include.
 export const UI_ISOLATED_TEST_FILES = [
   "ui/src/pages/chat/chat-pane-history.test.ts",
+  "ui/src/pages/chat/chat-pane-identity.test.ts",
   "ui/src/pages/chat/chat-pane-lifecycle.test.ts",
   "ui/src/pages/chat/chat-pane-pull-requests.test.ts",
   "ui/src/pages/chat/chat-pane.message-cut.test.ts",
