+++ ui/src/pages/chat/chat-pane.test.ts
@@ -168,37 +168,6 @@ describe("chat pane pull request refresh", () => {
     expect(pane.sessionPullRequests).toEqual([]);
   });
 
-  it("keeps a session-specific assistant identity across ordinary gateway snapshots", () => {
-    const client = {} as GatewayBrowserClient;
-    const { pane } = createTestChatPane({ client, sessions: {} as SessionCapability });
-    const state = (pane as unknown as { state: ChatPageHost }).state;
-    state.client = client;
-    state.connected = true;
-    state.assistantName = "Session Agent";
-
-    pane.applyGatewaySnapshot({
-      ...pane.context.gateway.snapshot,
-      client,
-    });
-
-    expect(state.assistantName).toBe("Session Agent");
-  });
-
-  it("resets a session-specific identity when the logical connection changes", () => {
-    const client = {} as GatewayBrowserClient;
-    const nextClient = {} as GatewayBrowserClient;
-    const { pane, state } = createTestChatPane({ client, sessions: {} as SessionCapability });
-    state.assistantName = "Session Agent";
-
-    pane.applyGatewaySnapshot({
-      ...pane.context.gateway.snapshot,
-      client: nextClient,
-      phase: "reconnecting" as const,
-    });
-
-    expect(state.assistantName).toBe(pane.context.config.current.assistantIdentity.name);
-  });
-
   it("preserves shared PR state for an empty rate-limited snapshot", async () => {
     const request = vi.fn().mockResolvedValue({ pullRequests: [], rateLimited: true });
     const setPullRequestSummary = vi.fn();
